requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# SAE and Interpretability
sae-lens>=0.1.0
einops>=0.6.0

# Data Processing and Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# NLP and Text Processing
nltk>=3.8.0
spacy>=3.6.0
pronouncing>=0.2.0
sentence-transformers>=2.2.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0
click>=8.1.0

# Optional: For advanced analysis
shap>=0.42.0
lime>=0.2.0 